auth improvements

+++ backend/src/data/mongo/user/functions.py
@@ -1,4 +1,4 @@
-from datetime import datetime, timedelta
+from datetime import datetime
 from typing import Any, Dict, Optional
 
 from src.models import UserPackage
@@ -7,21 +7,18 @@ from src.data.mongo.main import USERS
 from src.data.mongo.user.compression import compress
 
 
+async def is_user_key(user_id: str, user_key: str) -> bool:
+    user: Optional[dict[str, str]] = await USERS.find_one(  # type: ignore
+        {"user_id": user_id}, {"user_key": 1}
+    )
+    return user is not None and user["user_key"] == user_key
+
+
 async def lock_user(user_id: str) -> None:
     await USERS.update_one(  # type: ignore
         {"user_id": user_id},
         {"$set": {"lock": datetime.now()}},
     )
-
-
-async def is_user_locked(user_id: str) -> bool:
-    """Returns true if user is locked, false otherwise"""
-    user = await USERS.find_one({"user_id": user_id}, {"lock": 1})  # type: ignore
-    last_updated: datetime = datetime(1970, 1, 1)
-    if user is not None and user["lock"] is not None:
-        last_updated: datetime = user["lock"]
-    time_diff = datetime.now() - last_updated
-    return time_diff < timedelta(minutes=1)
 
 
 async def update_user_metadata(user_id: str, raw_user: Dict[str, Any]) -> None:
@@ -43,5 +40,8 @@ async def update_user_raw_data(
         )
 
 
-async def delete_user(user_id: str) -> None:
+async def delete_user(user_id: str, user_key: str, use_user_key: bool = True) -> bool:
+    if use_user_key and not is_user_key(user_id, user_key):
+        return False
     await USERS.delete_one({"user_id": user_id})  # type: ignore
+    return True
